requirements.txt
python-dotenv==1.1.1
discord==2.3.2
pytz==2025.2
google-cloud-firestore==2.21.0